requirements.txt
Django==3.2.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
ipython==7.26.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pillow==8.3.1
gunicorn==20.1.0